Move main ACL and Thing code into their folders
